1.Personal Assistant/requirements.txt
speechrecognition
pyttsx3
requests
pyaudio
python-dotenv